Refactor handleReadingStringValue method
